♻️ Rename the module about sync connection MCP client to be more clear.
