+++ test/integration_test/mcp_client/test_mcp_client_integration.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List
 
 import httpx
 
-from gearmeshing_ai.mcp_client.client import McpClient
+from gearmeshing_ai.mcp_client.sync import McpClient
 from gearmeshing_ai.mcp_client.schemas.config import (
     GatewayConfig,
     McpClientConfig,
