+++ test/unit_test/mcp_client/test_protocol_runtime_checks.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from gearmeshing_ai.mcp_client.client import McpClient
+from gearmeshing_ai.mcp_client.sync import McpClient
 from gearmeshing_ai.mcp_client.client_async import AsyncMcpClient
